SAY i CONNECT acabats!!!! (Per acabar fase dos queda exit i ficar maco SHOW CONNECTIONS)

+++ trinity/includes.h
@@ -44,7 +44,7 @@ typedef struct{
 typedef struct{
     char type;
     char* header;
-    int length;
+    short length;
     char* data;
 } Trama;
